Add monitor
Add the monitor abstraction.
